enable x64 support and fix 64-bit issues
git-svn-id: https://rtnet.svn.sourceforge.net/svnroot/rtnet/trunk@1106 be94bdbd-1fe7-0310-aeeb-c95d3c307607

+++ rtnet/examples/xenomai/native/tdma-api.c
@@ -84,6 +84,6 @@ int main(int argc, char *argv[])
         printf("cycle #%ld, start %.9f s, offset %lld ns\n",
                waitinfo.cycle_no,
                (waitinfo.cycle_start+waitinfo.clock_offset)/1000000000.0,
-               waitinfo.clock_offset);
+               (unsigned long long)waitinfo.clock_offset);
     }
 }
